requirements.txt
asgiref==3.8.1
dj-database-url==2.3.0
Django==5.2
django-cors-headers==4.7.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
pillow==11.1.0
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-decouple==3.8
python-dotenv==1.0.1
pytz==2025.1
sqlparse==0.5.3
typing_extensions==4.13.1
